requirements.txt
blinker>=1.9
bottle>=0.13,<0.14
build>=1.2
certifi>=2023
charset-normalizer>=3.3
click>=8.2
docopt>=0.6
Flask>=3.1
fastapi>=0.104.1
idna>=3.4
itsdangerous>=2.1
Jinja2>=3.1
joblib>=1.3
MarkupSafe>=2.1
numpy>=1.24,<2.0         # numpy 2.x aún tiene problemas en Windows
packaging>=23
pandas>=2.2
pipreqs>=0.4
proxy_tools>=0.1
psutil>=5.9               # 7.0 puede dar problemas en Windows
pyproject_hooks>=1.0
PyQt6>=6.5,<6.9.1         # 6.9.1 puede fallar sin compatibilidad binaria en Windows
PyQt6-WebEngine>=6.5,<6.9
PyQt6_sip>=13.5
python-dateutil>=2.8
pytz>=2023
pywebview>=4.0
QtPy>=2.2
requests>=2.30
scikit-learn>=1.2,<1.7    # 1.7 no siempre se instala bien en Windows sin compilador
scipy>=1.10,<1.15         # similar: versiones seguras sin compilar en Windows
setuptools>=65
six>=1.16
threadpoolctl>=3.1
typing_extensions>=4.0
tzdata>=2023
urllib3>=2.1
uvicorn[standard]>=0.24.0
Werkzeug>=2.3
wheel>=0.40
yarg>=0.1